+++ CHANGELOG.md
@@ -25,6 +25,8 @@
 -   Made `Game.current()` a classmethod instead of a property. This is to make it more obvious that it is read-only.
 -   Restructured the Joystick input system to make it easier to work with in more complex settings.
 -   Changed external `Time` methods and attributes to use seconds instead of milliseconds where applicable.
+-   Fullscreen option in `init()` is now a boolean.
+-   Added maximize option to `init()`.
 
 ### Added